a potential thread-racing fix
the version before was notifying that a particular PCH was built to
anyone waiting for it first and then was broadcasting that to
clangdserver which was potentially causing a re-parse of the opened
documents. Those documents, however, could be already in some
subsequent PCH-rebuild process, potentially accessing either an invalid
data or a data out of bounds from other threads while working with the
content of files

this commit first broadcasts the finished state of the item and then
changes the state and notifies the rest about it

cannot say that I have a total overview of side effects. I guess some
sort of a deadlock could happen since during broadcasting the actually finished
PCH is not marked as such and if there will be an attempt to wait for
readiness... well... we're screwed